requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.6
django-bootstrap4==23.2
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
easy-thumbnails==2.8.5
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
mysqlclient
